Rename package context->runtime
