+++ internal/pkg/cmd/application/list.go
@@ -21,11 +21,10 @@ import (
 	"github.com/spf13/cobra"
 	"github.com/wso2/choreo-cli/internal/pkg/client"
 	"github.com/wso2/choreo-cli/internal/pkg/cmd/common"
-	"github.com/wso2/choreo-cli/internal/pkg/cmd/context"
-	"github.com/wso2/choreo-cli/internal/pkg/config"
+	"github.com/wso2/choreo-cli/internal/pkg/cmd/runtime"
 )
 
-func NewListCommand(cliContext context.CliContext) *cobra.Command {
+func NewListCommand(cliContext runtime.CliContext) *cobra.Command {
 
 	const cmdList = "list"
 	cmd := &cobra.Command{
@@ -38,14 +37,14 @@ func NewListCommand(cliContext context.CliContext) *cobra.Command {
 	return cmd
 }
 
-func runListAppCommand(cliContext context.CliContext) func(cmd *cobra.Command, args []string) {
+func runListAppCommand(cliContext runtime.CliContext) func(cmd *cobra.Command, args []string) {
 	return func(cmd *cobra.Command, args []string) {
 
 		listApps(cliContext)
 	}
 }
 
-func listApps(cliContext context.CliContext) {
+func listApps(cliContext runtime.CliContext) {
 
 	req, err := client.NewRequest(cliContext, "GET", pathApplications, nil)
