+++ internal/pkg/cmd/runtime/context.go
@@ -7,7 +7,7 @@
  * You may not alter or remove any copyright or other notice from copies of this content.
  */
 
-package context
+package runtime
 
 import (
 	"io"
